requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.1
graphviz==0.20.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.7.3
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
psutil==5.9.5
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
threadpoolctl==3.2.0
tqdm==4.66.1
xgboost==2.0.0
